test(j2cl): harden mention helper parity cleanup
